requirements.txt
fastapi
uvicorn
PyMuPDF
tiktoken
sentence-transformers
faiss-cpu
openai
